+++ flow-typed/types/media-config-metadata.js
@@ -1,0 +1,5 @@
+// @flow
+declare type ProviderMediaConfigMetadataObject = {
+  name: string,
+  description: string
+};