+++ flow-typed/types/media-entry.js
@@ -6,5 +6,6 @@ declare type ProviderMediaEntryObject = {
   duration: number,
   dvrStatus: number,
   metadata: Object,
-  type: string
+  type: string,
+  poster: string | Array<Object>
 };