+++ src/k-provider/ott/config.js
@@ -2,8 +2,6 @@
 import {clone} from '../../util/clone'
 
 const defaultConfig: Object = {
-  serviceUrl: "//api-preprod.ott.kaltura.com/v4_6/api_v3",
-  cdnUrl: "//api-preprod.ott.kaltura.com/v4_7",
   serviceParams: {
     apiVersion: '4.7.1'
   }