+++ CHANGELOG.md
@@ -2,6 +2,62 @@
 
 All notable changes to this project will be documented in this file. See [standard-version](https://github.com/conventional-changelog/standard-version) for commit guidelines.
 
+<a name="2.3.2"></a>
+## [2.3.2](https://github.com/kaltura/playkit-js-providers/compare/v2.3.1...v2.3.2) (2018-05-26)
+
+
+### Bug Fixes
+
+* **FEC-8235:** remove default service url from OTT provider and analytics ([#54](https://github.com/kaltura/playkit-js-providers/issues/54)) ([43fb839](https://github.com/kaltura/playkit-js-providers/commit/43fb839))
+
+
+
+<a name="2.3.1"></a>
+## [2.3.1](https://github.com/kaltura/playkit-js-providers/compare/v2.3.0...v2.3.1) (2018-05-16)
+
+
+### Bug Fixes
+
+* **FEC-8217:** add asset reference for the asset get service ([#53](https://github.com/kaltura/playkit-js-providers/issues/53)) ([fe7ae71](https://github.com/kaltura/playkit-js-providers/commit/fe7ae71))
+
+
+
+<a name="2.3.0"></a>
+# [2.3.0](https://github.com/kaltura/playkit-js-providers/compare/v2.2.0...v2.3.0) (2018-05-06)
+
+
+### Bug Fixes
+
+* **FEC-8096:** some of OTT metadata is getting overriden ([#50](https://github.com/kaltura/playkit-js-providers/issues/50)) ([30a7553](https://github.com/kaltura/playkit-js-providers/commit/30a7553))
+* **FEC-8135:** change the provider default service config to https://cdsnapisec.kaltura.com ([#51](https://github.com/kaltura/playkit-js-providers/issues/51)) ([f05784b](https://github.com/kaltura/playkit-js-providers/commit/f05784b))
+
+
+### Features
+
+* **FEC-8176:** restructure media config ([#52](https://github.com/kaltura/playkit-js-providers/issues/52)) ([911d79d](https://github.com/kaltura/playkit-js-providers/commit/911d79d))
+
+
+
+<a name="2.2.0"></a>
+# [2.2.0](https://github.com/kaltura/playkit-js-providers/compare/v2.1.2...v2.2.0) (2018-03-12)
+
+
+### Features
+
+* **ott-provider:** retrieve poster from ott backend ([#43](https://github.com/kaltura/playkit-js-providers/issues/43)) ([bdd5da1](https://github.com/kaltura/playkit-js-providers/commit/bdd5da1))
+
+
+
+<a name="2.1.2"></a>
+## [2.1.2](https://github.com/kaltura/playkit-js-providers/compare/v2.1.1...v2.1.2) (2018-02-14)
+
+
+### Bug Fixes
+
+* **FEC-7828:** Live videos in OTT show VOD UI ([#49](https://github.com/kaltura/playkit-js-providers/issues/49)) ([d19c46a](https://github.com/kaltura/playkit-js-providers/commit/d19c46a))
+
+
+
 <a name="2.1.1"></a>
 ## [2.1.1](https://github.com/kaltura/playkit-js-providers/compare/v2.1.0...v2.1.1) (2018-02-11)